Cput

+++ src/3dmaskdump.c
@@ -15,15 +15,17 @@ int main( int argc , char * argv[] )
    int narg , nvox , ii,jj,kk,vv , mcount , iv , mc , ndset,ndval , i,j,k ;
    THD_3dim_dataset *mask_dset=NULL , **input_dset=NULL ;
    float mask_bot=666.0 , mask_top=-666.0 ;
-   byte * mmm   = NULL ;
-   char * oname = NULL , * obuf , * otemp ;
-   FILE * ofile ;
-   MRI_IMAGE * flim ;
-   float * flar ;
-   int noijk=0 , yes_xyz=0 ;
+   byte *mmm   = NULL ;
+   char *oname = NULL , *obuf , *otemp ;
+   FILE *ofile ;
+   MRI_IMAGE *flim ;
+   float *flar ;
+   int no_ijk=0 , yes_xyz=0 ;
    int yes_index=0 ;                    /*-- 09 May 2003 [rickr] --*/
    byte *cmask=NULL ; int ncmask=0 ;
    int verb=1 ;
+   int yes_niml=0 , no_zero=0 , numz ;   /* 04 Feb 2008 */
+   int nout ; char **bout , *niml_name="maskdump" ;
 
 #define BOXLEN   7
 #define BOX_XYZ  1
@@ -58,7 +60,7 @@ int main( int argc , char * argv[] )
              "  -xyz         Means to write the x,y,z coordinates from\n"
              "                 the 1st input dataset at the start of each\n"
              "                 output line.  These coordinates are in\n"
-             "                 the 'RAI' order.\n"
+             "                 the 'RAI' (DICOM) order.\n"
              "  -o fname     Means to write output to file 'fname'.\n"
              "                 [default = stdout, which you won't like]\n"
              "\n"
@@ -131,6 +133,17 @@ int main( int argc , char * argv[] )
              "              * Coordinates (for -xbox, -dbox, and -nbox) are relative\n"
              "                  to the first dataset on the command line.\n"
              "\n"
+             "  -nozero       Means to skip output of any voxel where all the\n"
+             "                  data values are zero.\n"
+             "\n"
+             "  -niml name    Means to output data in the XML/NIML format that\n"
+             "                  is compatible with input back to AFNI via\n"
+             "                  the READ_NIML_FILE command.\n"
+             "              * 'name' is the 'target_name' for the NIML header\n"
+             "                  field, which is the name that will be assigned\n"
+             "                  to the dataset when it is sent into AFNI.\n"
+             "              * Also implies '-noijk' and '-xyz' and '-nozero'.\n"
+             "\n"
              "  -quiet        Means not to print progress messages to stderr.\n"
              "\n"
              "Inputs after the last option are datasets whose values you\n"
@@ -145,6 +158,9 @@ int main( int argc , char * argv[] )
              "the mask value to be output, you have to include that\n"
              "dataset in the dataset input list again, after you use\n"
              "it in the '-mask' option.\n"
+             "\n"
+             "* To eliminate the 'i j k' columns, use the '-noijk' option.\n"
+             "* To add spatial coordinate columns, use the '-xyz' option.\n"
              "\n"
              "N.B.: This program doesn't work with complex-valued datasets!\n"
             ) ;
@@ -170,6 +186,20 @@ int main( int argc , char * argv[] )
    narg = 1 ;
    while( narg < argc && argv[narg][0] == '-' ){
 
+      if( strcasecmp(argv[narg],"-niml") == 0 ){             /* 04 Feb 2008 */
+        yes_niml = 1 ; narg++ ;
+        if( narg >= argc ) ERROR_exit("-niml must be followed by a name") ;
+        niml_name = strdup(argv[narg]) ;
+        if( !THD_filename_pure(niml_name) )
+          ERROR_exit("Illegal 'name' after '-niml'") ;
+        narg++ ; continue ;
+      }
+
+      if( strcasecmp(argv[narg],"-nozero") == 0 ){           /* 04 Feb 2008 */
+        no_zero = 1 ; narg++ ; continue ;
+      }
+
+
       if( strcmp(argv[narg],"-quiet") == 0 ){                /* 09 May 2003 - RWC */
         verb = 0 ; narg++ ; continue ;
       }
@@ -183,30 +213,25 @@ int main( int argc , char * argv[] )
           case 'd': btyp = BOX_DIC ; break ;
           case 'n': btyp = BOX_NEU ; break ;
           case 'i': btyp = BOX_IJK ; break ;
-          default:
-            fprintf(stderr,"** Unknown 'box' option %s\n",argv[narg]); exit(1);
+          default:  ERROR_exit("Unknown 'box' option %s\n",argv[narg]) ;
         }
-        if( narg+3 >= argc ){
-          fprintf(stderr,"** need 3 arguments after %s\n",argv[narg]); exit(1);
-        }
+        if( narg+3 >= argc )
+          ERROR_exit("need 3 arguments after %s\n",argv[narg]);
         nn = sscanf( argv[narg+1] , "%f:%f" , &xbot , &xtop ) ;
-        if( nn < 1 ){
-          fprintf(stderr,"** Can't decode %s after %s\n",argv[narg+1],argv[narg]); exit(1);
-        } else if( nn == 1 ){
+        if( nn < 1 )
+          ERROR_exit("Can't decode %s after %s\n",argv[narg+1],argv[narg]);
+        else if( nn == 1 )
           xtop=xbot ;
-        }
         nn = sscanf( argv[narg+2] , "%f:%f" , &ybot , &ytop ) ;
-        if( nn < 1 ){
-          fprintf(stderr,"** Can't decode %s after %s\n",argv[narg+2],argv[narg]); exit(1);
-        } else if( nn == 1 ){
+        if( nn < 1 )
+          ERROR_exit("Can't decode %s after %s\n",argv[narg+2],argv[narg]);
+        else if( nn == 1 )
           ytop=ybot ;
-        }
         nn = sscanf( argv[narg+3] , "%f:%f" , &zbot , &ztop ) ;
-        if( nn < 1 ){
-          fprintf(stderr,"** Can't decode %s after %s\n",argv[narg+3],argv[narg]); exit(1);
-        } else if( nn == 1 ){
+        if( nn < 1 )
+          ERROR_exit("Can't decode %s after %s\n",argv[narg+3],argv[narg]);
+        else if( nn == 1 )
           ztop=zbot ;
-        }
         box_dat = (float *) realloc( box_dat , sizeof(float)*BOXLEN*(box_num+1) ) ;
         box_dat[0+BOXLEN*box_num] = xbot ;
         box_dat[1+BOXLEN*box_num] = xtop ;
@@ -225,7 +250,7 @@ int main( int argc , char * argv[] )
       }
 
       if( strcmp(argv[narg],"-noijk") == 0 ){
-         noijk = 1 ;
+         no_ijk = 1 ;
          narg++ ; continue ;
       }
 
@@ -235,64 +260,52 @@ int main( int argc , char * argv[] )
       }
 
       if( strcmp(argv[narg],"-cmask") == 0 ){  /* 16 Mar 2000 */
-         if( narg+1 >= argc ){
-            fprintf(stderr,"** -cmask option requires a following argument!\n");
-            exit(1) ;
-         }
+         if( narg+1 >= argc )
+            ERROR_exit("-cmask option requires a following argument!\n");
          cmask = EDT_calcmask( argv[++narg] , &ncmask, 0 ) ;
-         if( cmask == NULL ){
-            fprintf(stderr,"** Can't compute -cmask!\n"); exit(1);
-         }
+         if( cmask == NULL ) ERROR_exit("Can't compute -cmask!\n");
          narg++ ; continue ;
       }
 
       if( strncmp(argv[narg],"-mask",5) == 0 ){
-         if( mask_dset != NULL ){
-            fprintf(stderr,"** Cannot have two -mask options!\n") ; exit(1) ;
-         }
-         if( narg+1 >= argc ){
-            fprintf(stderr,"** -mask option requires a following argument!\n");
-            exit(1) ;
-         }
+         if( mask_dset != NULL )
+           ERROR_exit("Cannot have two -mask options!\n") ;
+         if( narg+1 >= argc )
+           ERROR_exit("-mask option requires a following argument!\n");
          mask_dset = THD_open_dataset( argv[++narg] ) ;
-         if( mask_dset == NULL ){
-            fprintf(stderr,"** Cannot open mask dataset!\n") ; exit(1) ;
-         }
-         if( DSET_BRICK_TYPE(mask_dset,0) == MRI_complex ){
-            fprintf(stderr,"** Cannot deal with complex-valued mask dataset!\n");
-            exit(1) ;
-         }
+         if( mask_dset == NULL )
+           ERROR_exit("Cannot open mask dataset!\n") ;
+         if( DSET_BRICK_TYPE(mask_dset,0) == MRI_complex )
+           ERROR_exit("Cannot deal with complex-valued mask dataset!\n");
          narg++ ; continue ;
       }
 
       if( strncmp(argv[narg],"-mrange",5) == 0 ){
-         if( narg+2 >= argc ){
-            fprintf(stderr,"** -mrange option requires 2 following arguments!\n");
-             exit(1) ;
-         }
+         if( narg+2 >= argc )
+           ERROR_exit("-mrange option requires 2 following arguments!\n");
          mask_bot = strtod( argv[++narg] , NULL ) ;
          mask_top = strtod( argv[++narg] , NULL ) ;
-         if( mask_top < mask_top ){
-            fprintf(stderr,"** -mrange inputs are illegal!\n") ; exit(1) ;
-         }
+         if( mask_top < mask_top )
+           ERROR_exit("-mrange inputs are illegal!\n") ;
          narg++ ; continue ;
       }
 
       if( strcmp(argv[narg],"-o") == 0 ){
-         if( narg+1 >= argc ){
-            fprintf(stderr,"** -o needs an argument after it!\n"); exit(1);
-         }
+         if( narg+1 >= argc )
+           ERROR_exit("-o needs an argument after it!\n");
          oname = argv[++narg] ;
-         if( ! THD_filename_ok(oname) ){
-            fprintf(stderr,"** name after -o is illegal!\n"); exit(1) ;
-         }
-         if( THD_is_file(oname) ){
-            fprintf(stderr,"** file %s already exists!\n",oname); exit(1);
-         }
+         if( ! THD_filename_ok(oname) )
+           ERROR_exit("name after -o is illegal!\n"); 
+         if( THD_is_file(oname) )
+           ERROR_exit("file %s already exists!\n",oname);
          narg++ ; continue ;
       }
 
-      fprintf(stderr,"** Unknown option: %s\n",argv[narg]) ; exit(1) ;
+      ERROR_exit("Unknown option: %s\n",argv[narg]) ;
+   }
+
+   if( yes_niml ){
+     no_ijk = 1; yes_xyz = 1; yes_index = 0; no_zero = 1; /* 04 Feb 2008 */
    }
 
    /* keep -quiet quiet                      27 Mar 2006 [rickr] */
@@ -301,35 +314,27 @@ int main( int argc , char * argv[] )
    /* should have at least one more argument */
 
    ndset = argc - narg ;
-   if( ndset <= 0 ){
-      fprintf(stderr,"** No input dataset!?\n") ; exit(1) ;
-   }
+   if( ndset <= 0 ) ERROR_exit("No input dataset!?\n") ;
 
    /* open all input datasets */
 
    input_dset = (THD_3dim_dataset **) malloc( sizeof(THD_3dim_dataset *)*ndset ) ;
    for( ndval=ii=0 ; ii < ndset ; ii++ ){
       input_dset[ii] = THD_open_dataset( argv[narg+ii] ) ;
-      if( input_dset[ii] == NULL ){
-         fprintf(stderr,"** Can't open dataset %s!\n",argv[narg+ii]) ;
-         exit(1) ;
-      }
+      if( input_dset[ii] == NULL )
+        ERROR_exit("Can't open dataset %s!\n",argv[narg+ii]) ;
 
-      if( DSET_BRICK_TYPE(input_dset[ii],0) == MRI_complex ){
-         fprintf(stderr,"** Cannot deal with complex-valued input dataset!\n");
-         exit(1) ;
-      }
+      if( DSET_BRICK_TYPE(input_dset[ii],0) == MRI_complex )
+        ERROR_exit("Cannot deal with complex-valued input dataset!\n");
 
       if( ii == 0 ){
-         nvox = DSET_NVOX(input_dset[0]) ;
-         nx   = DSET_NX(input_dset[0]) ;
-         ny   = DSET_NY(input_dset[0]) ; nxy  = nx*ny ;
-         nz   = DSET_NZ(input_dset[0]) ; nxyz = nxy*nz ;
-      } else {
-         if( DSET_NVOX(input_dset[ii]) != nvox ){
-            fprintf(stderr,"** Dataset %s does not match in size!\n",argv[narg+ii]);
-            exit(1) ;
-         }
+        nvox = DSET_NVOX(input_dset[0]) ;
+        nx   = DSET_NX(input_dset[0]) ;
+        ny   = DSET_NY(input_dset[0]) ; nxy  = nx*ny ;
+        nz   = DSET_NZ(input_dset[0]) ; nxyz = nxy*nz ;
+      } else if( DSET_NX(input_dset[ii]) != nx ||
+                 DSET_NY(input_dset[ii]) != ny || DSET_NZ(input_dset[ii]) != nz ){
+        ERROR_exit("Dataset %s does not match in size!\n",argv[narg+ii]);
       }
 
       ndval += DSET_NVALS(input_dset[ii]) ;
@@ -339,44 +344,33 @@ int main( int argc , char * argv[] )
 
    if( mask_dset == NULL ){
       mmm = NULL ;
-      if( verb ) fprintf(stderr,"++ %d voxels in the entire dataset (no mask)\n",nvox) ;
+      if( verb ) INFO_message("%d voxels in the entire dataset (no mask)\n",nvox) ;
    } else {
-      if( DSET_NVOX(mask_dset) != nvox ){
-         fprintf(stderr,"** Input and mask datasets are not same dimensions!\n");
-         exit(1) ;
-      }
+      if( DSET_NVOX(mask_dset) != nvox )
+        ERROR_exit("Input and mask datasets are not same dimensions!\n");
       mmm = THD_makemask( mask_dset , 0 , mask_bot,mask_top ) ;
       mcount = THD_countmask( nvox , mmm ) ;
-      if( mcount <= 0 ){
-         fprintf(stderr,"** No voxels in the mask!\n") ; exit(1) ;
-      }
-      if( verb ) fprintf(stderr,"++ %d voxels in the mask\n",mcount) ;
+      if( mcount <= 0 ) ERROR_exit("No voxels in the mask!\n") ;
+      if( verb ) INFO_message("%d voxels in the mask\n",mcount) ;
       DSET_delete(mask_dset) ;
    }
 
    /* 16 Mar 2000: deal with the cmask */
 
    if( cmask != NULL ){
-      if( ncmask != nvox ){
-         fprintf(stderr,"** Input and cmask datasets are not same dimensions!\n");
-         exit(1) ;
-      }
+      if( ncmask != nvox )
+        ERROR_exit("Input and cmask datasets are not same dimensions!\n");
       if( mmm != NULL ){
-         for( ii=0 ; ii < nvox ; ii++ )
-            mmm[ii] = (mmm[ii] && cmask[ii]) ;
+         for( ii=0 ; ii < nvox ; ii++ ) mmm[ii] = (mmm[ii] && cmask[ii]) ;
          free(cmask) ;
          mcount = THD_countmask( nvox , mmm ) ;
-         if( mcount <= 0 ){
-            fprintf(stderr,"** No voxels in the mask+cmask!\n") ; exit(1) ;
-         }
-         if( verb ) fprintf(stderr,"++ %d voxels in the mask+cmask\n",mcount) ;
+         if( mcount <= 0 ) ERROR_exit("No voxels in the mask+cmask!\n") ;
+         if( verb ) INFO_message("%d voxels in the mask+cmask\n",mcount) ;
       } else {
          mmm = cmask ;
          mcount = THD_countmask( nvox , mmm ) ;
-         if( mcount <= 0 ){
-            fprintf(stderr,"** No voxels in the cmask!\n") ; exit(1) ;
-         }
-         if( verb ) fprintf(stderr,"++ %d voxels in the cmask\n",mcount) ;
+         if( mcount <= 0 ) ERROR_exit("No voxels in the cmask!\n") ;
+         if( verb ) INFO_message("%d voxels in the cmask\n",mcount) ;
       }
    }
 
@@ -442,24 +436,20 @@ int main( int argc , char * argv[] )
 
        for( kk=kbot ; kk <= ktop ; kk++ )
         for( jj=jbot ; jj <= jtop ; jj++ )
-         for( ii=ibot ; ii <= itop ; ii++ )
-           bmask[ii+jj*nx+kk*nxy] = 1 ;
+         for( ii=ibot ; ii <= itop ; ii++ ) bmask[ii+jj*nx+kk*nxy] = 1 ;
      }
 
      mcount = THD_countmask( nvox , bmask ) ;
-     if( verb ) fprintf(stderr,"++ %d voxels in the boxes\n",mcount) ;
-     if( mcount == 0 ){
-       fprintf(stderr,"** Can't continue with no voxels insides boxes!\n"); exit(1);
-     }
+     if( verb ) INFO_message("%d voxels in the boxes\n",mcount) ;
+     if( mcount == 0 )
+       ERROR_exit("Can't continue with no voxels insides boxes!\n");
      if( mmm != NULL ){
        for( ii=0 ; ii < nvox ; ii++ )
           mmm[ii] = (mmm[ii] && bmask[ii]) ;
        free(bmask) ;
        mcount = THD_countmask( nvox , mmm ) ;
-       if( mcount <= 0 ){
-          fprintf(stderr,"** No voxels in the mask+boxes!\n") ; exit(1) ;
-       }
-       if( verb ) fprintf(stderr,"++ %d voxels in the mask+boxes\n",mcount) ;
+       if( mcount <= 0 ) ERROR_exit("No voxels in the mask+boxes!\n") ;
+       if( verb ) INFO_message("%d voxels in the mask+boxes\n",mcount) ;
      } else {
        mmm = bmask ;
      }
@@ -468,18 +458,16 @@ int main( int argc , char * argv[] )
    /* read in input dataset bricks */
 
    for( ii=0 ; ii < ndset ; ii++ ){
-      DSET_load(input_dset[ii]) ;  CHECK_LOAD_ERROR(input_dset[ii]) ;
+     DSET_load(input_dset[ii]) ;  CHECK_LOAD_ERROR(input_dset[ii]) ;
    }
 
    /* open the output file */
 
    if( oname != NULL ){
-      ofile = fopen( oname , "w" ) ;
-      if( ofile == NULL ){
-         fprintf(stderr,"*** Can't open output file %s\n",oname); exit(1);
-      }
+     ofile = fopen( oname , "w" ) ;
+     if( ofile == NULL ) ERROR_exit("Can't open output file %s\n",oname);
    } else {
-      ofile = stdout ;
+     ofile = stdout ;
    }
 
    /* output string buffers */
@@ -489,25 +477,26 @@ int main( int argc , char * argv[] )
 
    /* loop over voxels */
 
+   nout = 0 ; bout = NULL ;
    for( ii=0 ; ii < nvox ; ii++ ){
       if( mmm != NULL && mmm[ii] == 0 ) continue ;  /* skip this voxel */
 
-      obuf[0] = '\0' ;
+      obuf[0] = '\0' ; /* start it off with nothing */
 
       /*-- 09 May 2003: optionally display index    [rickr] --*/
       if ( yes_index ){
-         otemp = MV_format_fval((float)ii);
-         strcat(obuf,otemp); strcat(obuf," ");
+        otemp = MV_format_fval((float)ii);
+        strcat(obuf,otemp); strcat(obuf," ");
       }
 
-      if( ! noijk ){
-         i = DSET_index_to_ix( input_dset[0] , ii ) ;  /* voxel indexes */
-         j = DSET_index_to_jy( input_dset[0] , ii ) ;
-         k = DSET_index_to_kz( input_dset[0] , ii ) ;
+      if( ! no_ijk ){
+        i = DSET_index_to_ix( input_dset[0] , ii ) ;  /* voxel indexes */
+        j = DSET_index_to_jy( input_dset[0] , ii ) ;
+        k = DSET_index_to_kz( input_dset[0] , ii ) ;
 
-         otemp = MV_format_fval((float)i); strcat(obuf,otemp); strcat(obuf," ");
-         otemp = MV_format_fval((float)j); strcat(obuf,otemp); strcat(obuf," ");
-         otemp = MV_format_fval((float)k); strcat(obuf,otemp); strcat(obuf," ");
+        otemp = MV_format_fval((float)i); strcat(obuf,otemp); strcat(obuf," ");
+        otemp = MV_format_fval((float)j); strcat(obuf,otemp); strcat(obuf," ");
+        otemp = MV_format_fval((float)k); strcat(obuf,otemp); strcat(obuf," ");
       }
 
       if( yes_xyz ){                                  /* 23 Mar 2003 */
@@ -523,22 +512,50 @@ int main( int argc , char * argv[] )
         otemp = MV_format_fval(vec.xyz[2]); strcat(obuf,otemp); strcat(obuf," ");
       }
 
-      for( kk=0 ; kk < ndset ; kk++ ){
-         flim = THD_extract_series( ii , input_dset[kk] , 0 ) ;
-         flar = MRI_FLOAT_PTR(flim) ;
-         for( vv=0 ; vv < flim->nx ; vv++ ){
-            otemp = MV_format_fval(flar[vv]) ;
-            strcat(obuf,otemp) ; strcat(obuf," ") ;
-         }
-         mri_free(flim) ;
+      for( numz=kk=0 ; kk < ndset ; kk++ ){
+        flim = THD_extract_series( ii , input_dset[kk] , 0 ) ;
+        flar = MRI_FLOAT_PTR(flim) ;
+        for( vv=0 ; vv < flim->nx ; vv++ ){
+           if( flar[vv] == 0.0f ) numz++ ;         /* 04 Feb 2008 */
+           otemp = MV_format_fval(flar[vv]) ;
+           strcat(obuf,otemp) ; strcat(obuf," ") ;
+        }
+        mri_free(flim) ;
       }
+      if( no_zero && numz == ndval ) continue ;     /* 04 Feb 2008 */
 
       jj = strlen(obuf) ; obuf[jj-1] = '\0' ; /* kill last blank */
 
-      fprintf(ofile,"%s\n",obuf) ;
+      if( yes_niml ){           /* 04 Feb 2008: save output for NIML-ization */
+        nout++ ;
+        bout = (char **)realloc(bout,sizeof(char *)*nout) ;
+        bout[nout-1] = strdup(obuf) ;
+      } else {
+        fprintf(ofile,"%s\n",obuf) ;  /* regular output */
+      }
+
+   } /* end of loop over voxels */
+
+   if( yes_niml ){  /*----- 04 Feb 2008: NIML formatted output -----*/
+     if( nout > 0 ){
+       fprintf(ofile,"<VOLUME_DATA_SPARSE\n") ;
+       fprintf(ofile,"  master_name='%s'\n",DSET_IDCODE_STR(input_dset[0]) ) ;
+       fprintf(ofile,"  target_name='%s'\n",niml_name) ;
+       fprintf(ofile,"  ni_type='%d*float'\n",ndval+3) ; /* +3 allows for xyz */
+       fprintf(ofile,"  ni_dimen='%d'\n",nout) ;
+       fprintf(ofile,">\n") ;
+       for( ii=0 ; ii < nout ; ii++ ){
+         fprintf(ofile,"%s\n",bout[ii]); free(bout[ii]);
+       }
+       fprintf(ofile,"</VOLUME_DATA_SPARSE>\n") ;
+       free(bout) ;
+       if( verb ) INFO_message("-niml: output %d voxels",nout) ;
+     } else {
+       ERROR_exit("-niml output requested, but no data to output!") ;
+     }
    }
 
-   fflush(ofile) ; fsync(fileno(ofile)) ; fclose(ofile) ;
-   exit(0) ;
-}
+   /*----- Free at last! -----*/
 
+   fflush(ofile); fsync(fileno(ofile)); fclose(ofile); exit(0) ;
+}
